+++ mitol-django-digital-credentials/mitol/digitalcredentials/views_test.py
@@ -1,3 +1,0 @@
-"""Views tests"""
-# Views depend on models requiring an installed app, so the tests are instead in:
-#   mitol-django-digitalcredentials/testapp/views_test.py