+++ mitol-django-digital-credentials/testapp/models.py
@@ -1,4 +1,5 @@
 """Testapp models"""
+from django.conf import settings
 from django.db import models
 
 
@@ -7,3 +8,5 @@ class DemoCourseware(models.Model):
 
     title = models.CharField(max_length=100)
     description = models.TextField()
+
+    learner = models.ForeignKey(settings.AUTH_USER_MODEL, on_delete=models.CASCADE)